Make the train_model able to loop over the whole dataset.
Parameter dimensions changed to take only a single theta1, theta2, beta, and
alpha value (since they are the same for all the gates anyways). An updated
variational circuit is created everytime the objective function is called. The
referential circuit is updated when the input data is changed (the expected
result is also updated).